Basic version ready

+++ backend/bots/examples/group_prize_bot/api/src/updates/initalize_bot.rs
@@ -1,18 +1,18 @@
 use candid::CandidType;
 use serde::{Deserialize, Serialize};
 use std::fmt::Debug;
-use types::{Cryptocurrency, TimestampMillis, CanisterId, Cycles};
+use types::{CanisterId, Cryptocurrency, Cycles, TimestampMillis};
 
 #[derive(CandidType, Serialize, Deserialize, Debug)]
 pub struct Args {
     pub username: String,
     pub token: Cryptocurrency,
     pub ledger_canister_id: CanisterId,
-    pub max_individual_prize: u128,
-    pub min_individual_prize: u128,
+    pub max_individual_prize: u64,
+    pub min_individual_prize: u64,
     pub min_claimants_per_message: u32,
     pub max_claimants_per_message: u32,
-    pub end_date: TimestampMillis,    
+    pub end_date: TimestampMillis,
     pub groups: Vec<CanisterId>,
 }
